requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
rembg==2.0.50
Pillow==10.0.0
Werkzeug==3.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
